requirements.txt
flask
Flask-SQLAlchemy
pytest
python-decouple
psycopg2
alembic
python-dotenv